Revert "Rotas do gerenciamento da ementa (CRUD ementa)"
