+++ package.json
@@ -7,9 +7,7 @@
     "test": "jest --runInBand",
     "dev": "npx nodemon ./src/server.js",
     "build": "npx sequelize-cli db:migrate",
-    "start": "node ./src/server.js",
-    "seed":"npx sequelize-cli db:seed:all",
-    "unseed": "npx sequelize-cli db:seed:undo:all"
+    "start": "node ./src/server.js"
   },
   "repository": {
     "type": "git",
