--- seeders/20210210065604-create-course.js
@@ -1,17 +1,0 @@
-/* eslint-disable no-unused-vars */
-module.exports = {
-  up: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkInsert('courses', [{
-      title: 'JavaScript do zero ao avançado',
-      description: 'Curso para vc ficar voando mesmo tipo mostrão no JS',
-      color: '#e2ef08',
-      imageUrl: 'https://i.imgur.com/lWUs38z.png',
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    }], {});
-  },
-
-  down: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkDelete('courses', null, {});
-  },
-};
