+++ src/controllers/coursesController.js
@@ -1,10 +1,5 @@
 /* eslint-disable class-methods-use-this */
 const Course = require('../models/Course');
-const Chapter = require('../models/Chapter');
-const Topic = require('../models/Topic');
-const TopicUser = require('../models/TopicUser');
-const User = require('../models/User');
-const CourseUser = require('../models/CourseUser');
 const ConflictError = require('../errors/ConflictError');
 const NotFoundError = require('../errors/NotFoundError');
 
@@ -14,26 +9,8 @@ class CoursesController {
   }
 
   async getOne(id) {
-    let course = await Course.findByPk(id, {
-      attributes: {
-        exclude: ['createdAt', 'updatedAt'],
-      },
-      order: [[{ model: Chapter }, 'order', 'ASC']],
-      include: {
-        model: Chapter,
-        attributes: {
-          exclude: ['courseId', 'order', 'createdAt', 'updatedAt'],
-        },
-      },
-    });
-
+    const course = await Course.findByPk(id);
     if (!course) throw new NotFoundError('Course not found');
-
-    course = course.toJSON();
-
-    let totalTopicsQuantity = 0;
-    course.chapters.forEach((c) => { totalTopicsQuantity += c.topicsQuantity; });
-    course = { totalTopicsQuantity, ...course };
 
     return course;
   }
@@ -42,9 +19,11 @@ class CoursesController {
     const course = await Course.findOne({ where: { title } });
     if (course !== null) throw new ConflictError('There is already a course with this title');
 
-    return Course.create({
+    const createdCourse = await Course.create({
       title, description, color, imageUrl,
     });
+
+    return createdCourse;
   }
 
   async editCourse(id, title, description, color, imageUrl) {
@@ -70,35 +49,6 @@ class CoursesController {
 
   count() {
     return Course.count();
-  }
-
-  async initCouserByUserId(courseId, userId) {
-    const course = await Course.findByPk(courseId);
-    const user = await User.findByPk(userId);
-
-    if (!course || !user) throw new NotFoundError();
-    return CourseUser.create({ courseId, userId });
-  }
-
-  getAllTopicsAtChapterFromUser(courseId, chapterId, userId) {
-    return Chapter.findByPk(chapterId, {
-      where: { courseId },
-      attributes: {
-        exclude: ['createdAt', 'updatedAt', 'courseId'],
-      },
-      include: {
-        model: Topic,
-        attributes: {
-          exclude: ['createdAt', 'updatedAt'],
-        },
-        include: {
-          model: TopicUser,
-          attributes: ['userId'],
-          where: { userId },
-          required: false,
-        },
-      },
-    });
   }
 }
