+++ src/middlewares/validation.js
@@ -7,8 +7,6 @@ async function verifyJWT(req, res, next) {
   try {
     let token = req.header('Authorization');
 
-    if (!token) return res.status(400).send({ message: 'Token not found' });
-
     const split = token.split(' ');
     token = split[1];
