--- src/models/CourseUser.js
@@ -1,41 +1,0 @@
-const { Sequelize } = require('sequelize');
-const sequelize = require('../utils/database');
-
-class CourseUser extends Sequelize.Model { }
-
-CourseUser.init(
-  {
-    id: {
-      type: Sequelize.INTEGER,
-      primaryKey: true,
-      autoIncrement: true,
-      allowNull: false,
-    },
-    userId: {
-      type: Sequelize.INTEGER,
-      allowNull: false,
-      references: {
-        model: 'users',
-        key: 'id',
-      },
-    },
-    courseId: {
-      type: Sequelize.INTEGER,
-      allowNull: false,
-      references: {
-        model: 'courses',
-        key: 'id',
-      },
-    },
-    doneActivities: {
-      type: Sequelize.INTEGER,
-      defaultValue: 0,
-    },
-  },
-  {
-    sequelize,
-    modelName: 'courseUser',
-  },
-);
-
-module.exports = CourseUser;
