--- src/models/TopicUser.js
@@ -1,37 +1,0 @@
-const { Sequelize } = require('sequelize');
-const sequelize = require('../utils/database');
-
-class TopicUser extends Sequelize.Model { }
-
-TopicUser.init(
-  {
-    id: {
-      type: Sequelize.INTEGER,
-      primaryKey: true,
-      autoIncrement: true,
-      allowNull: false,
-    },
-    topicId: {
-      type: Sequelize.INTEGER,
-      allowNull: false,
-      references: {
-        model: 'topics',
-        key: 'id',
-      },
-    },
-    userId: {
-      type: Sequelize.INTEGER,
-      allowNull: false,
-      references: {
-        model: 'users',
-        key: 'id',
-      },
-    },
-  },
-  {
-    sequelize,
-    modelName: 'topicUser',
-  },
-);
-
-module.exports = TopicUser;
