+++ src/routers/usersRouter.js
@@ -3,23 +3,25 @@ const usersController = require('../controllers/usersController');
 const registerSchema = require('../schemas/registerSchema');
 const signInSchema = require('../schemas/signInSchema');
 
-router.post('/register', async (req, res) => {
-  const { error } = registerSchema.validate(req.body);
-  if (error) return res.status(422).send({ error: error.details[0].message });
-  if (await usersController.findByEmail(req.body.email)) return res.status(409).send({ error: 'Email already in use' });
+router
+  .post('/register', async (req, res) => {
+    const { error } = registerSchema.validate(req.body);
+    if (error) return res.status(422).send({ error: error.details[0].message });
+    if (await usersController.findByEmail(req.body.email)) return res.status(409).send({ error: 'Email already in use' });
 
-  const user = await usersController.create(req.body);
-  return res.status(201).send(user);
-});
+    const user = await usersController.create(req.body);
+    return res.status(201).send(user);
+  });
 
-router.post('/sign-in', async (req, res) => {
-  const { error } = signInSchema.validate(req.body);
-  if (error) return res.status(422).send({ error: error.details[0].message });
+router
+  .post('/sign-in', async (req, res) => {
+    const { error } = signInSchema.validate(req.body);
+    if (error) return res.status(422).send({ error: error.details[0].message });
 
-  const { email, password } = req.body;
-  const session = await usersController.createSession(email, password);
+    const { email, password } = req.body;
+    const session = await usersController.createSession(email, password);
 
-  return res.status(201).send(session);
-});
+    return res.status(201).send(session);
+  });
 
 module.exports = router;
