+++ tests/integration/app.test.js
@@ -11,26 +11,15 @@ const db = new Pool({
   connectionString: process.env.DATABASE_URL,
 });
 
-async function cleanDatabase() {
-  await db.query('DELETE FROM "theoryUsers"');
-  await db.query('DELETE FROM "topicUsers"');
-  await db.query('DELETE FROM "exerciseUsers"');
-  await db.query('DELETE FROM theories');
-  await db.query('DELETE FROM exercises');
-  await db.query('DELETE FROM topics');
-  await db.query('DELETE FROM chapters');
-  await db.query('DELETE FROM "courseUsers"');
-  await db.query('DELETE FROM "adminSessions"');
+beforeEach(async () => {
   await db.query('DELETE FROM sessions');
-  await db.query('DELETE FROM courses');
   await db.query('DELETE FROM users');
-  await db.query('ALTER SEQUENCE courses_id_seq RESTART WITH 1;');
-}
-
-beforeAll(cleanDatabase);
+});
 
 afterAll(async () => {
-  await cleanDatabase();
+  await db.query('DELETE FROM sessions');
+  await db.query('DELETE FROM users');
+
   await db.end();
   await sequelize.close();
 });
